# resources/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Restaurant Reservation API
  description: API for managing restaurant reservations, orders, and payments.
  version: 1.0.0
servers:
  - url: http://localhost/api
    description: Local development server
paths:
...
  /check-availability:
    get:
      summary: Check table availability
      operationId: checkAvailability
      tags:
        - Reservations
      security:
        - bearerAuth: []
      parameters:
        - name: customer_id
          in: query
          required: true
          schema:
            type: integer
            example: 1
        - name: from_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: 2024-08-10 17:00:00
        - name: to_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: 2024-08-10 19:00:00
        - name: guests
          in: query
          required: true
          schema:
            type: integer
            example: 4
      responses:
        '200':
          description: Availability checked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  available_tables:
                    type: array
                    items:
                      type: integer
                      example: 1
        '400':
          description: Invalid request
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT